Update pricingscript.r
the score on test data is 0.344980. It is a random forest model with 200 trees, 5 nodesize and no CV.

currently running the randomforest in caret package with build in CV settings. 10 fold, and 1 repeats. if the performance increases, maybe consider to increase the number of repeats to 5,10,15,and 20

Next step plan, initially i remove the product_id column from the training data set, because i found that the test data include new products that are not in the training set, considering adding the variable back to the model

Just read that it seems many people include the specs file in to the model, will try to add the spec variables in the model too.